+++ src/continuity/discrete/__init__.py
@@ -1,0 +1,10 @@
+"""
+`continuity.discrete`
+
+Functionalities handling discretization of continuous functionals.
+"""
+
+from .uniform import UniformBoxSampler
+from .regular_grid import RegularGridSampler
+
+__all__ = ["UniformBoxSampler", "RegularGridSampler"]