+++ src/continuity/operators/common.py
@@ -75,59 +75,20 @@ class DeepResidualNetwork(torch.nn.Module):
         depth: int,
         act: Optional[torch.nn.Module] = None,
     ):
+        assert depth >= 1, "DeepResidualNetwork has at least depth 1."
         super().__init__()
 
         self.act = act or torch.nn.Tanh()
         self.first_layer = torch.nn.Linear(input_size, width)
         self.hidden_layers = torch.nn.ModuleList(
-            [ResidualLayer(width, act=self.act) for _ in range(depth)]
+            [ResidualLayer(width, act=self.act) for _ in range(1, depth)]
         )
         self.last_layer = torch.nn.Linear(width, output_size)
 
     def forward(self, x):
         """Forward pass."""
         x = self.first_layer(x)
+        x = self.act(x)
         for layer in self.hidden_layers:
             x = layer(x)
         return self.last_layer(x)
-
-
-class NeuralNetworkKernel(torch.nn.Module):
-    """Neural network kernel.
-
-    Args:
-        kernel_width: Width of kernel network
-        kernel_depth: Depth of kernel network
-
-    TODO: As it is implemented, this is a convolution `k(x - y)`, but it should be more general.
-    """
-
-    def __init__(
-        self,
-        kernel_width: int,
-        kernel_depth: int,
-    ):
-        super().__init__()
-        self.net = DeepResidualNetwork(
-            1,
-            1,
-            kernel_width,
-            kernel_depth,
-        )
-
-    def forward(self, x: torch.Tensor, y: torch.Tensor) -> torch.Tensor:
-        """Compute kernel value.
-
-        Args:
-            x: Tensor of shape (..., coordinate_dim)
-            y: Tensor of shape (..., coordinate_dim)
-
-        Returns:
-            Tensor of shape (...)
-        """
-        r = ((x - y) ** 2).sum(dim=-1)
-        output_shape = r.shape
-        r = r.reshape((-1, 1))
-        k = self.net(r)
-        k = k.reshape(output_shape)
-        return k