+++ src/continuity/trainer/trainer.py
@@ -3,12 +3,19 @@
 """
 
 import torch
+import torch.distributed as dist
+from torch.nn.parallel import DistributedDataParallel as DDP
+from torch.utils.data import DataLoader
+from torch.utils.data.distributed import DistributedSampler
 from time import time
 from typing import Optional, List
+from continuity.data import OperatorDataset
 from continuity.operators import Operator
 from continuity.operators.losses import Loss, MSELoss
-from continuity.trainer.callbacks import Callback, PrintTrainingLoss
 from continuity.trainer.device import get_device
+from .callbacks import Callback, PrintTrainingLoss
+from .criterion import Criterion, TrainingLossCriterion
+from .logs import Logs
 
 
 class Trainer:
@@ -22,12 +29,12 @@ class Trainer:
         optimizer = torch.optim.Adam(operator.parameters(), lr=1e-3)
         loss_fn = MSELoss()
         trainer = Trainer(operator, optimizer, loss_fn, device="cuda:0")
-        trainer.fit(data_loader, epochs=100)
+        trainer.fit(dataset, tol=1e-3, epochs=1000)
         ```
 
     Args:
         operator: Operator to be trained.
-        optimizer: Torch-like optimizer. Default is Adam.
+        optimizer: Torch-like optimizer. Default is Adam with learning rate 1e-3.
         loss_fn: Loss function taking (op, x, u, y, v). Default is MSELoss.
         device: Device to train on. Default is CPU.
         verbose: Print model parameters and use PrintTrainingLoss callback by default. Default is True.
@@ -41,7 +48,7 @@ class Trainer:
         optimizer: Optional[torch.optim.Optimizer] = None,
         loss_fn: Optional[Loss] = None,
         device: torch.device = device,
-        verbose: bool = True,
+        verbose: Optional[bool] = None,
     ):
         self.operator = operator
         self.optimizer = (
@@ -51,20 +58,29 @@ class Trainer:
         )
         self.loss_fn = loss_fn if loss_fn is not None else MSELoss()
         self.device = device
-        self.verbose = verbose
+        self.rank = device.index or 0
+        self.verbose = verbose if verbose is not None else self.rank == 0
 
     def fit(
         self,
-        data_loader: torch.utils.data.DataLoader,
-        epochs: int = 100,
+        dataset: OperatorDataset,
+        tol: float = 1e-5,
+        epochs: int = 1000,
         callbacks: Optional[List[Callback]] = None,
+        criterion: Optional[Criterion] = None,
+        batch_size: int = 32,
+        shuffle: bool = True,
     ):
         """Fit operator to data set.
 
         Args:
             dataset: Data set.
-            epochs: Number of epochs.
-            callbacks: List of callbacks.
+            tol: Tolerance for stopping criterion. Ignored if criterion is not None.
+            epochs: Maximum number of epochs.
+            callbacks: List of callbacks. Defaults to [PrintTrainingLoss] if verbose.
+            criterion: Stopping criterion. Defaults to TrainingLossCriteria(tol).
+            batch_size: Batch size.
+            shuffle: Shuffle data set.
         """
         # Default callback
         if callbacks is None:
@@ -73,20 +89,45 @@ class Trainer:
             else:
                 callbacks = []
 
+        # Default criterion
+        if criterion is None:
+            criterion = TrainingLossCriterion(tol)
+
         # Print number of model parameters
         if self.verbose:
             num_params = sum(p.numel() for p in self.operator.parameters())
             print(f"Model parameters: {num_params}")
 
         # Move operator to device
-        self.operator.to(self.device)
+        operator = self.operator.to(self.device)
+
+        # Use DistributedDataParallel if available
+        sampler = None
+        if dist.is_available() and dist.is_initialized():
+            operator = DDP(
+                operator, device_ids=[self.device], output_device=self.device
+            )
+            sampler = DistributedSampler(dataset)
+            shuffle = False
+
+            if self.verbose:
+                ngpu = dist.get_world_size()
+                print(f"Device: CUDA ({ngpu} GPU{'' if ngpu == 1 else 's'})")
+        else:
+            if self.verbose:
+                print(f"Device: {self.device}")
+
+        # Create data loader
+        data_loader = DataLoader(
+            dataset, batch_size=batch_size, shuffle=shuffle, sampler=sampler
+        )
 
         # Call on_train_begin
         for callback in callbacks:
             callback.on_train_begin()
 
         # Train
-        self.operator.train()
+        operator.train()
         for epoch in range(epochs):
             loss_train = 0
 
@@ -97,27 +138,33 @@ class Trainer:
 
                 def closure(x=x, u=u, y=y, v=v):
                     self.optimizer.zero_grad()
-                    loss = self.loss_fn(self.operator, x, u, y, v)
+                    loss = self.loss_fn(operator, x, u, y, v)
                     loss.backward(retain_graph=True)
                     return loss
 
                 self.optimizer.step(closure)
 
                 # Compute mean loss
-                loss_train += self.loss_fn(self.operator, x, u, y, v).detach().item()
+                loss_train += self.loss_fn(operator, x, u, y, v).detach().item()
 
             end = time()
             seconds_per_epoch = end - start
             loss_train /= len(data_loader)
 
             # Callbacks
-            logs = {
-                "loss/train": loss_train,
-                "seconds_per_epoch": seconds_per_epoch,
-            }
+            logs = Logs(
+                epoch=epoch + 1,
+                loss_train=loss_train,
+                seconds_per_epoch=seconds_per_epoch,
+            )
 
             for callback in callbacks:
-                callback(epoch + 1, logs)
+                callback(logs)
+
+            # Stopping criterion
+            if criterion is not None:
+                if criterion(logs):
+                    break
 
         # Call on_train_end
         for callback in callbacks: