+++ tests/operators/test_belnet.py
@@ -1,9 +1,6 @@
 import torch
-import matplotlib.pyplot as plt
 import pytest
 
-from continuity.plotting import plot, plot_evaluation
-from torch.utils.data import DataLoader
 from continuity.operators import BelNet
 from continuity.data import OperatorDataset
 from continuity.data.sine import OperatorDataset, Sine
@@ -51,28 +48,15 @@ def test_belnet():
 
     # Data set
     dataset = Sine(num_sensors, size=1)
-    data_loader = DataLoader(dataset, batch_size=1, shuffle=True)
 
     # Operator
-    operator = BelNet(
-        dataset.shapes,
-    )
+    operator = BelNet(dataset.shapes)
 
-    # Train self-supervised
-    optimizer = torch.optim.Adam(operator.parameters(), lr=1e-3)
-    trainer = Trainer(operator, optimizer)
-    trainer.fit(data_loader, epochs=1000)
-
-    # Plotting
-    fig, ax = plt.subplots(1, 1)
-    x, u, _, _ = dataset[0]
-    plot(x, u, ax=ax)
-    plot_evaluation(operator, x, u, ax=ax)
-    fig.savefig(f"test_belnet.png")
+    # Train
+    Trainer(operator).fit(dataset, tol=1e-3, batch_size=1)
 
     # Check solution
-    x = x.unsqueeze(0)
-    u = u.unsqueeze(0)
+    x, u = dataset.x, dataset.u
     assert MSELoss()(operator, x, u, x, u) < 1e-3